# contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "commercial_v1_ListProvisioningTasksResponse": {
        "properties": {
          "nextAfterTaskId": {
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_ProvisioningTaskDTO"
            },
            "type": "array"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ListProvisioningTasksResponse"
      },
...
      "commercial_v1_MutateProvisioningTaskRequest": {
        "properties": {
          "expectedRevision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "reason": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.MutateProvisioningTaskRequest"
      },
...
      "commercial_v1_ProvisioningCompletionDTO": {
        "properties": {
          "appliedAt": {
            "type": "string"
          },
          "changeId": {
            "type": "string"
          },
          "entitlementVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "sourceVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "subscriptionRevision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ProvisioningCompletionDTO"
      },
...
      "commercial_v1_ProvisioningRequirementDTO": {
        "properties": {
          "adapter": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "maxAttempts": {
            "minimum": 0,
            "type": "integer"
          },
          "version": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ProvisioningRequirementDTO"
      },
      "commercial_v1_ProvisioningStepDTO": {
        "properties": {
          "attempts": {
            "minimum": 0,
            "type": "integer"
          },
          "cycleAttempts": {
            "minimum": 0,
            "type": "integer"
          },
          "effect": {
            "type": "string"
          },
          "evidence": {
            "type": "string"
          },
          "failureCode": {
            "type": "string"
          },
          "reconciliations": {
            "minimum": 0,
            "type": "integer"
          },
          "requirement": {
            "$ref": "#/components/schemas/commercial_v1_ProvisioningRequirementDTO"
          },
          "state": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ProvisioningStepDTO"
      },
      "commercial_v1_ProvisioningTaskDTO": {
        "properties": {
          "actorId": {
            "type": "string"
          },
          "cancellationAllowed": {
            "type": "boolean"
          },
          "changeId": {
            "type": "string"
          },
          "completion": {
            "$ref": "#/components/schemas/commercial_v1_ProvisioningCompletionDTO"
          },
          "createdAt": {
            "type": "string"
          },
          "deadline": {
            "type": "string"
          },
          "failureCode": {
            "type": "string"
          },
          "nextAttemptAt": {
            "type": "string"
          },
          "retryAllowed": {
            "type": "boolean"
          },
          "retryCycles": {
            "minimum": 0,
            "type": "integer"
          },
          "revision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "stage": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "stepIndex": {
            "minimum": 0,
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_ProvisioningStepDTO"
            },
            "type": "array"
          },
          "targetPlanCode": {
            "type": "string"
          },
          "targetPlanVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "taskId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ProvisioningTaskDTO"
      },
...
    }
  },
  "info": {
    "title": "yunka API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/platform/tenants/{tenant_id}/provisioning/tasks": {
      "get": {
        "operationId": "commercial.v1.ProvisioningApplication.ListProvisioningTasks",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "afterTaskId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_ListProvisioningTasksResponse"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.ProvisioningApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.ProvisioningApplication/ListProvisioningTasks"
      }
    },
    "/v1/platform/tenants/{tenant_id}/provisioning/tasks/{task_id}": {
      "get": {
        "operationId": "commercial.v1.ProvisioningApplication.GetProvisioningTask",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_ProvisioningTaskDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.ProvisioningApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.ProvisioningApplication/GetProvisioningTask"
      }
    },
    "/v1/platform/tenants/{tenant_id}/provisioning/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "commercial.v1.ProvisioningApplication.CancelProvisioningTask",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commercial_v1_MutateProvisioningTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_ProvisioningTaskDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.ProvisioningApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.ProvisioningApplication/CancelProvisioningTask"
      }
    },
    "/v1/platform/tenants/{tenant_id}/provisioning/tasks/{task_id}/retry": {
      "post": {
        "operationId": "commercial.v1.ProvisioningApplication.RetryProvisioningTask",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commercial_v1_MutateProvisioningTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_ProvisioningTaskDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.ProvisioningApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.ProvisioningApplication/RetryProvisioningTask"
      }
    },
...
  },
...
}