# internal/commercial/transport/rest/zz_yunka_provisioning_operation_executor_gen.go
// Code generated by yunka contract; DO NOT EDIT.

package rest

import (
	"errors"
	commercialv1 "github.com/hvritual/biz/contracts/gen/commercial/v1"
	application "github.com/hvritual/biz/internal/commercial/application"
	policy "github.com/hvritual/biz/internal/commercial/policy"
	codes "google.golang.org/grpc/codes"
	status "google.golang.org/grpc/status"
	protojson "google.golang.org/protobuf/encoding/protojson"
	io "io"
	"net/http"
	strconv "strconv"
	execution "yunka.io/framework/execution"
	operation "yunka.io/framework/operation"
	authz "yunka.io/gateway/authz"
	httpbinding "yunka.io/gateway/httpbinding"
)

type ProvisioningOperationHandler struct {
	application application.ProvisioningApplication
	executor    operation.Executor
}

func RegisterProvisioningOperationExecutor(mux *http.ServeMux, application application.ProvisioningApplication, executor operation.Executor) error {
	if mux == nil {
		return errors.New("contract C9 REST adapter: mux is required")
	}
	if application == nil {
		return errors.New("contract C9 REST adapter: application is required")
	}
	if executor == nil {
		return errors.New("contract C9 REST adapter: operation executor is required")
	}
	handler := &ProvisioningOperationHandler{application: application, executor: executor}
	if err := httpbinding.Register(mux, "POST", "/v1/platform/tenants/{tenant_id}/provisioning/tasks/{task_id}/cancel", handler.handleOperationCancelProvisioningTask); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/platform/tenants/{tenant_id}/provisioning/tasks/{task_id}", handler.handleOperationGetProvisioningTask); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/platform/tenants/{tenant_id}/provisioning/deliveries", handler.handleOperationListProvisioningDeliveries); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/platform/tenants/{tenant_id}/provisioning/tasks", handler.handleOperationListProvisioningTasks); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "POST", "/v1/platform/tenants/{tenant_id}/provisioning/tasks/{task_id}/retry", handler.handleOperationRetryProvisioningTask); err != nil {
		return err
	}
	return nil
}

func writeProvisioningOperationError(writer http.ResponseWriter, err error) {
	if authz.IsDenied(err) {
		statusCode := http.StatusForbidden
		var denied *authz.DeniedError
		if errors.As(err, &denied) && (denied.Decision.Reason == authz.ReasonUnauthenticated || denied.Decision.Reason == authz.ReasonAuthenticationMethod) {
			statusCode = http.StatusUnauthorized
		}
		http.Error(writer, http.StatusText(statusCode), statusCode)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyKeyRequired) {
		http.Error(writer, "idempotency key required", http.StatusBadRequest)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyInProgress) || errors.Is(err, execution.ErrIdempotencyCompleted) {
		http.Error(writer, "idempotency conflict", http.StatusConflict)
		return
	}
	if code := status.Code(err); code == codes.Aborted || code == codes.AlreadyExists {
		http.Error(writer, "application conflict", http.StatusConflict)
		return
	}
	if errors.Is(err, operation.ErrExecutorUnavailable) || errors.Is(err, operation.ErrSecurityUnavailable) || errors.Is(err, operation.ErrSecurityNilContext) || errors.Is(err, operation.ErrIdempotencyUnavailable) {
		http.Error(writer, "operation execution unavailable", http.StatusInternalServerError)
		return
	}
	http.Error(writer, "application request failed", http.StatusBadRequest)
}

func (handler *ProvisioningOperationHandler) handleOperationCancelProvisioningTask(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.MutateProvisioningTaskRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.TenantId = request.PathValue("tenant_id")
	wire.TaskId = request.PathValue("task_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanProvisioningCancelProvisioningTask(), wire, handler.application.CancelProvisioningTask)
	if err != nil {
		writeProvisioningOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *ProvisioningOperationHandler) handleOperationGetProvisioningTask(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.ReadProvisioningTaskRequest{}
	wire.TenantId = request.PathValue("tenant_id")
	wire.TaskId = request.PathValue("task_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanProvisioningGetProvisioningTask(), wire, handler.application.GetProvisioningTask)
	if err != nil {
		writeProvisioningOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *ProvisioningOperationHandler) handleOperationListProvisioningDeliveries(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.ListProvisioningDeliveriesRequest{}
	if raw := request.URL.Query().Get("after_event_id"); raw != "" {
		wire.AfterEventId = raw
	}
	if raw := request.URL.Query().Get("limit"); raw != "" {
		parsed, err := strconv.ParseUint(raw, 10, 32)
		if err != nil {
			http.Error(writer, "invalid request parameter", http.StatusBadRequest)
			return
		}
		wire.Limit = uint32(parsed)
	}
	wire.TenantId = request.PathValue("tenant_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanProvisioningListProvisioningDeliveries(), wire, handler.application.ListProvisioningDeliveries)
	if err != nil {
		writeProvisioningOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *ProvisioningOperationHandler) handleOperationListProvisioningTasks(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.ListProvisioningTasksRequest{}
	if raw := request.URL.Query().Get("after_task_id"); raw != "" {
		wire.AfterTaskId = raw
	}
	if raw := request.URL.Query().Get("limit"); raw != "" {
		parsed, err := strconv.ParseUint(raw, 10, 32)
		if err != nil {
			http.Error(writer, "invalid request parameter", http.StatusBadRequest)
			return
		}
		wire.Limit = uint32(parsed)
	}
	wire.TenantId = request.PathValue("tenant_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanProvisioningListProvisioningTasks(), wire, handler.application.ListProvisioningTasks)
	if err != nil {
		writeProvisioningOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *ProvisioningOperationHandler) handleOperationRetryProvisioningTask(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.MutateProvisioningTaskRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.TenantId = request.PathValue("tenant_id")
	wire.TaskId = request.PathValue("task_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanProvisioningRetryProvisioningTask(), wire, handler.application.RetryProvisioningTask)
	if err != nil {
		writeProvisioningOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}